requirements.txt
fastapi
uvicorn
chromadb
pypdf
pdfplumber
python-dotenv
openai
pydantic
pydantic-settings
aiofiles
jinja2
pillow
numpy
pandas
matplotlib
seaborn
scikit-learn
reportlab
pypdf2
qrcode[pil]